requirements.txt
# Core ML/NLP - Compatible versions
transformers==4.40.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
sentence-transformers==2.6.1
huggingface-hub==0.22.2
tokenizers==0.19.1
safetensors==0.4.2
accelerate==0.29.3

# NLP
spacy==3.7.4
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl

# Text extraction
pdfplumber==0.11.0
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8

# Utilities
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
pyyaml==6.0.1
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
nltk==3.8.1